pyproject.toml
[project]
name = "portal27"
version = "0.1.0"
description = "Image generation helper scripts for Portal 27"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-generativeai",
  "google-cloud-aiplatform",
  "openai",
  "Pillow",
  "python-dotenv",
  "requests",
  "beautifulsoup4>=4.14.3",
]

[tool.uv]
# Default UV settings; adjust if you need custom indexes or cache dirs.
